Pre: Add a Dumpable trait for lower level debugging of datom stores.
These functions are really useful when writing tests and quickly
iterating on code, but I'm pretty sure we don't want to expose this
outside of test configurations.  They're very low level, they're not
efficient, and they're definitely not idiomatic.  I don't think
there's a good way expose index walking to consumers in a similar way
to Datomic, which is what the efficient and idiomatic way to express
these debugging functions would be.