Part 4: Support tempid equivalence.
This completes the transition to Datomic-like input-as-set semantics,
which allows us to address #532.  Previously, two tempids that
upserted to the same entid would produce duplicate datoms, and that
would have been rejected by the transactor -- correctly, since we did
not allow duplicate datoms under the input-as-list semantics.  With
input-as-set semantics, duplicate datoms are allowed; and that means
that we must allow tempids to be equivalent, i.e., to resolve to the
same tempid.

To achieve this, we:
- index the set of tempids
- identify tempid indices that share an upsert
- map tempids to a dense set of contiguous integer labels

We use the well-known union-find algorithm, as implemented by
petgraph, to efficiently manage the set of equivalent tempids.