Pre: Sort debug::Datoms in Rust, not SQL.
This makes `Datoms` own more of its own sorting and display (as
`edn::Value`).  The goal is to make it pleasant to create
`debug::Datoms` instances that don't originate from an SQL query; the
immediate consumer is a `TransactWatcher` that collects witnessed
datoms... into a `debug::Datoms` instance, ready to compare with
`assert_matches!`.